nambahin HotelFilter dan update controller

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\ApiResponses;
 use App\Filters\UserFilter;
+use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
 
@@ -14,21 +15,25 @@ class UserController extends Controller
 
     public function index(UserFilter $filters)
     {
-        $query = (new UserFilter(request()))->apply(User::query());
-        $user = $query->paginate(10);
-        return $this->success($user, "Filter user success", 200);
+        $query = $filters->apply(User::query());
+        $users = $query->paginate(10);
+
+        return $this->success($users, "Filter user success", 200);
     }
 
-    public function show($id)
+    public function show(UserFilter $filters, $id)
     {
-        $user = User::find($id);
+        $query = $filters->apply(User::where("id", $id));
+        $user = $query->first();
+
         if (!$user) {
             return $this->error("User not found", 404);
         }
+
         return $this->success($user, "User found successfully", 200);
     }
 
-    public function store()
+    public function store(UserFilter $filters)
     {
         $user = User::create([
             'name' => request('name'),
@@ -36,24 +41,46 @@ class UserController extends Controller
             'password' => bcrypt(request('password'))
         ]);
 
-        return $this->success($user, "Create user success", 201);
+        if (request()->has('role')) {
+            $role = Role::where('slug', request('role'))->first();
+            if ($role) {
+                $user->roles()->attach($role->id);
+            }
+        }
+
+        return $this->success($user->load('roles'), "Create user success", 201);
     }
 
-    public function update($id)
+    public function update(UserFilter $filters, $id)
     {
         $user = User::find($id);
+        if (!$user) {
+            return $this->error("User not found", 404);
+        }
+
         $user->update([
             'name' => request('name'),
             'email' => request('email'),
         ]);
 
-        return $this->success($user, "User updated successfully", 201);
+        $query = $filters->apply(User::where("id", $user->id));
+        $user = $query->first();
+
+        return $this->success($user, "User updated successfully", 200);
     }
 
-    public function destroy($id)
+    public function destroy(UserFilter $filters, $id)
     {
-        $user = User::findOrFail($id);
+        $user = User::find($id);
+        if (!$user) {
+            return $this->error("User not found", 404);
+        }
+
         $user->delete();
-        return $this->success($user, "User deleted successfully", 200);
+
+        $query = $filters->apply(User::query());
+        $users = $query->paginate(10);
+
+        return $this->success($users, "User deleted successfully", 200);
     }
 }
